# merged-line-openapi.yaml
openapi: 3.0.0
info:
  title: Merged API
  version: 1.0.0
  description: Merged OpenAPI specification
paths:
...
  /v2/bot/message/multicast:
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-multicast-message
      tags:
      - messaging-api
      operationId: multicast
      description: An API that efficiently sends the same message to multiple user
        IDs. You can't send messages to group chats or multi-person chats.
      parameters:
      - name: X-Line-Retry-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: 'Retry key.

          Specifies the UUID in hexadecimal format (e.g., `123e4567-e89b-12d3-a456-426614174000`)
          generated by any method.

          The retry key isn''t generated by LINE. Each developer must generate their
          own retry key.

          '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MulticastRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MulticastResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v2/bot/message/narrowcast:
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-narrowcast-message
      tags:
      - messaging-api
      operationId: narrowcast
      description: Send narrowcast message
      parameters:
      - name: X-Line-Retry-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: 'Retry key.

          Specifies the UUID in hexadecimal format (e.g., `123e4567-e89b-12d3-a456-426614174000`)
          generated by any method.

          The retry key isn''t generated by LINE. Each developer must generate their
          own retry key.

          '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NarrowcastRequest'
        required: true
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrowcastResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v2/bot/message/broadcast:
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-broadcast-message
      tags:
      - messaging-api
      operationId: broadcast
      description: Sends a message to multiple users at any time.
      parameters:
      - name: X-Line-Retry-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: 'Retry key.

          Specifies the UUID in hexadecimal format (e.g., `123e4567-e89b-12d3-a456-426614174000`)
          generated by any method.

          The retry key isn''t generated by LINE. Each developer must generate their
          own retry key.

          '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BroadcastResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v2/bot/insight/followers:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-number-of-followers
      tags:
      - insight
      summary: Get number of followers
      description: 'Returns the number of users who have added the LINE Official Account

        on or before a specified date.

        '
      operationId: getNumberOfFollowers
      parameters:
      - in: query
        name: date
        required: false
        description: 'Date for which to retrieve the number of followers.


          Format: yyyyMMdd (e.g. 20191231)

          Timezone: UTC+9

          '
        schema:
          type: string
          pattern: ^[0-9]{8}$
          maxLength: 8
          minLength: 8
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetNumberOfFollowersResponse'
              example:
                status: ready
                followers: 7620
                targetedReaches: 5848
                blocks: 237
  /v2/bot/insight/message/event:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-message-event
      tags:
      - insight
      description: 'Returns statistics about how users interact with narrowcast messages

        or broadcast messages sent from your LINE Official Account.

        '
      summary: Get user interaction statistics
      operationId: getMessageEvent
      parameters:
      - in: query
        name: requestId
        required: true
        description: 'Request ID of a narrowcast message or broadcast message.

          Each Messaging API request has a request ID.

          '
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessageEventResponse'
              example:
                overview:
                  requestId: f70dd685-499a-4231-a441-f24b8d4fba21
                  timestamp: 1568214000
                  delivered: 320
                  uniqueImpression: 82
                  uniqueClick: 51
                  uniqueMediaPlayed: null
                  uniqueMediaPlayed100Percent: null
                messages:
                - seq: 1
                  impression: 136
                  mediaPlayed: null
                  mediaPlayed25Percent: null
                  mediaPlayed50Percent: null
                  mediaPlayed75Percent: null
                  mediaPlayed100Percent: null
                  uniqueMediaPlayed: null
                  uniqueMediaPlayed25Percent: null
                  uniqueMediaPlayed50Percent: null
                  uniqueMediaPlayed75Percent: null
                  uniqueMediaPlayed100Percent: null
                clicks:
                - seq: 1
                  url: https://line.me/
                  click: 41
                  uniqueClick: 30
                  uniqueClickOfRequest: 30
                - seq: 1
                  url: https://www.linebiz.com/
                  click: 59
                  uniqueClick: 38
                  uniqueClickOfRequest: 38
  /v2/bot/insight/message/event/aggregation:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-statistics-per-unit
      tags:
      - insight
      description: 'You can check the per-unit statistics of how users interact with

        push messages and multicast messages sent from your LINE Official Account.

        '
      operationId: getStatisticsPerUnit
      parameters:
      - in: query
        name: customAggregationUnit
        required: true
        description: 'Name of aggregation unit specified when sending the message.
          Case-sensitive.

          For example, `Promotion_a` and `Promotion_A` are regarded as different unit
          names.

          '
        schema:
          type: string
          pattern: ^[a-zA-Z0-9_]{1,30}$
          minLength: 1
          maxLength: 30
      - in: query
        name: from
        required: true
        description: 'Start date of aggregation period.


          Format: yyyyMMdd (e.g. 20210301)

          Time zone: UTC+9

          '
        schema:
          type: string
          pattern: ^[0-9]{8}$
          example: '20210301'
          minLength: 8
          maxLength: 8
      - in: query
        name: to
        required: true
        description: 'End date of aggregation period. The end date can be specified
          for up to 30 days later. For example, if the start date is 20210301, the
          latest end date is 20210331.


          Format: yyyyMMdd (e.g. 20210301)

          Time zone: UTC+9

          '
        schema:
          type: string
          pattern: ^[0-9]{8}$
          example: '20210301'
          minLength: 8
          maxLength: 8
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStatisticsPerUnitResponse'
              example:
                overview:
                  uniqueImpression: 40
                  uniqueClick: 30
                  uniqueMediaPlayed: 25
                  uniqueMediaPlayed100Percent: null
                messages:
                - seq: 1
                  impression: 42
                  mediaPlayed: 30
                  mediaPlayed25Percent: null
                  mediaPlayed50Percent: null
                  mediaPlayed75Percent: null
                  mediaPlayed100Percent: null
                  uniqueMediaPlayed: 25
                  uniqueMediaPlayed25Percent: null
                  uniqueMediaPlayed50Percent: null
                  uniqueMediaPlayed75Percent: null
                  uniqueMediaPlayed100Percent: null
                clicks:
                - seq: 1
                  url: https://developers.line.biz/
                  click: 35
                  uniqueClick: 25
                  uniqueClickOfRequest: null
                - seq: 1
                  url: https://www.line-community.me/
                  click: 29
                  uniqueClick: null
                  uniqueClickOfRequest: null
...
components:
  schemas:
...
    SentMessage:
      required:
      - id
      type: object
      properties:
        id:
          type: string
          description: ID of the sent message.
        quoteToken:
          type: string
          description: 'Quote token of the message. Only included when a message object
            that can be specified as a quote target was sent as a push or reply message.

            '
...
    NotificationDisabled:
      type: boolean
      default: false
      description: "`true`: The user doesn\u2019t receive a push notification when\
        \ a message is sent.\n`false`: The user receives a push notification when\
        \ the message is sent (unless they have disabled push notifications in LINE\
        \ and/or their device).\nThe default value is false.\n"
...
    MulticastRequest:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-multicast-message
      required:
      - messages
      - to
      type: object
      properties:
        messages:
          maxItems: 5
          minItems: 1
          type: array
          description: Messages to send
          items:
            $ref: '#/components/schemas/Message'
        to:
          maxItems: 500
          minItems: 1
          type: array
          description: Array of user IDs. Use userId values which are returned in
            webhook event objects. Do not use LINE IDs found on LINE.
          items:
            type: string
        notificationDisabled:
          $ref: '#/components/schemas/NotificationDisabled'
        customAggregationUnits:
          type: array
          maxItems: 1
          description: Name of aggregation unit. Case-sensitive.
          items:
            type: string
            maxLength: 30
            minLength: 1
            pattern: ^[a-zA-Z0-9_]{1,30}$
    MulticastResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-multicast-response
      type: object
    NarrowcastRequest:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-narrowcast-message
      required:
      - messages
      type: object
      properties:
        messages:
          maxItems: 5
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: List of Message objects.
        recipient:
          $ref: '#/components/schemas/Recipient'
        filter:
          $ref: '#/components/schemas/Filter'
        limit:
          $ref: '#/components/schemas/Limit'
        notificationDisabled:
          $ref: '#/components/schemas/NotificationDisabled'
    NarrowcastResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-narrowcast-response
      type: object
    Recipient:
      type: object
      description: Recipient
      properties:
        type:
          description: Type of recipient
          type: string
      discriminator:
        propertyName: type
        mapping:
          operator: '#/components/schemas/OperatorRecipient'
          audience: '#/components/schemas/AudienceRecipient'
          redelivery: '#/components/schemas/RedeliveryRecipient'
...
    AudienceRecipient:
      type: object
      allOf:
      - $ref: '#/components/schemas/Recipient'
      - type: object
        properties:
          audienceGroupId:
            type: integer
            format: int64
...
    Filter:
      type: object
      description: Filter for narrowcast
      properties:
        demographic:
          $ref: '#/components/schemas/DemographicFilter'
...
    Limit:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-narrowcast-message
      type: object
      description: Limit of the Narrowcast
      properties:
        max:
          minimum: 1
          type: integer
          format: int32
          description: 'The maximum number of narrowcast messages to send.

            Use this parameter to limit the number of narrowcast messages sent.

            The recipients will be chosen at random.

            '
        upToRemainingQuota:
          type: boolean
          default: false
          description: 'If true, the message will be sent within the maximum number
            of deliverable messages. The default value is `false`.


            Targets will be selected at random.

            '
...
    BroadcastRequest:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-broadcast-message
      required:
      - messages
      type: object
      properties:
        messages:
          maxItems: 5
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: List of Message objects.
        notificationDisabled:
          $ref: '#/components/schemas/NotificationDisabled'
    BroadcastResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#send-broadcast-response
      type: object
...
    Message:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#message-common-properties
      required:
      - type
      type: object
      properties:
        type:
          type: string
          description: Type of message
        quickReply:
          $ref: '#/components/schemas/QuickReply'
        sender:
          $ref: '#/components/schemas/Sender'
      discriminator:
        propertyName: type
        mapping:
          text: '#/components/schemas/TextMessage'
          textV2: '#/components/schemas/TextMessageV2'
          sticker: '#/components/schemas/StickerMessage'
          image: '#/components/schemas/ImageMessage'
          video: '#/components/schemas/VideoMessage'
          audio: '#/components/schemas/AudioMessage'
          location: '#/components/schemas/LocationMessage'
          imagemap: '#/components/schemas/ImagemapMessage'
          template: '#/components/schemas/TemplateMessage'
          flex: '#/components/schemas/FlexMessage'
...
    ErrorResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#error-responses
      type: object
      required:
      - message
      properties:
        message:
          type: string
          description: Message containing information about the error.
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: An array of error details. If the array is empty, this property
            will not be included in the response.
        sentMessages:
          description: Array of sent messages.
          maxItems: 5
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SentMessage'
    ErrorDetail:
      type: object
      properties:
        message:
          type: string
          description: Details of the error. Not included in the response under certain
            situations.
        property:
          type: string
          description: Location of where the error occurred. Returns the JSON field
            name or query parameter name of the request. Not included in the response
            under certain situations.
...
    GetNumberOfFollowersResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-number-of-followers
      description: Get number of followers
      type: object
      properties:
        status:
          description: Calculation status.
          type: string
          enum:
          - ready
          - unready
          - out_of_service
        followers:
          type: integer
          format: int64
          description: 'The number of times, as of the specified date, that a user
            added this LINE Official Account as a friend for the first time.

            The number doesn''t decrease even if a user later blocks the account or
            when they delete their LINE account.

            '
        targetedReaches:
          format: int64
          type: integer
          description: 'The number of users, as of the specified date, that the LINE
            Official Account can reach through targeted messages based on gender,
            age, and/or region.

            This number only includes users who are active on LINE or LINE services
            and whose demographics have a high level of certainty.

            '
        blocks:
          type: integer
          format: int64
          description: 'The number of users blocking the account as of the specified
            date.

            The number decreases when a user unblocks the account.



            '
    GetMessageEventResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-insight-message-event-response
      description: Statistics about how users interact with narrowcast messages or
        broadcast messages sent from your LINE Official Account.
      properties:
        overview:
          $ref: '#/components/schemas/GetMessageEventResponseOverview'
        messages:
          description: Array of information about individual message bubbles.
          type: array
          items:
            $ref: '#/components/schemas/GetMessageEventResponseMessage'
        clicks:
          description: Array of information about opened URLs in the message.
          type: array
          items:
            $ref: '#/components/schemas/GetMessageEventResponseClick'
      type: object
...
    GetMessageEventResponseClick:
      type: object
      properties:
        seq:
          type: integer
          format: int32
          description: The URL's serial number.
        url:
          type: string
          description: URL.
        click:
          type: integer
          format: int64
          description: Number of times the URL was opened.
          nullable: true
        uniqueClick:
          type: integer
          format: int64
          description: Number of users that opened the URL.
          nullable: true
        uniqueClickOfRequest:
          type: integer
          format: int64
          description: Number of users who opened this url through any link in the
            message. If a message contains two links to the same URL and a user opens
            both links, they're counted only once.
          nullable: true
    GetStatisticsPerUnitResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-statistics-per-unit-response
      description: Response object for `get statistics per unit`
      type: object
      properties:
        overview:
          $ref: '#/components/schemas/GetStatisticsPerUnitResponseOverview'
        messages:
          description: Array of information about individual message bubbles.
          type: array
          items:
            $ref: '#/components/schemas/GetStatisticsPerUnitResponseMessage'
        clicks:
          description: Array of information about opened URLs in the message.
          type: array
          items:
            $ref: '#/components/schemas/GetStatisticsPerUnitResponseClick'
      required:
      - clicks
      - messages
      - overview
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Channel access token
    basicAuth:
      type: http
      scheme: basic